% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kknn.R
\name{kknn_func}
\alias{kknn_func}
\title{Return Prediction of Sale Price}
\usage{
kknn_func(input)
}
\arguments{
\item{input}{Input taken from a shiny user interface}
}
\value{
A predicted value of sales price
}
\description{
This fucntion creates a predicted sale price using variables input through a UI.
It does this by creating a KKNN predictive model from a data set imported from a connected SQL database.
}
\examples{
kknn(input)
}
