% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kknn.R
\name{server}
\alias{server}
\title{Takes Input from the User Interface and Return a Value for the User}
\usage{
server(input, output)
}
\arguments{
\item{input}{Input taken from a shiny user interface}

\item{output}{Variable to be displayed to the user in the UI}
}
\value{
A value to user interface
}
\description{
This fucntion takes some input from a user interface built in shiny and returns a value for the user.
This value is calculated user a separate function called inside this function.
}
\examples{
server(input, output)
}
